chore(tests): restore LocaleSwitcher cache + drop stale eslint-disables
Two pre-existing drifts on main were blocking the Wave 32 Task 1 pre-flight
(100% test pass + <=7 lint warnings):

1. `LocaleSwitcher.test.tsx > persists the chosen locale...` had been failing
   since Wave 31 shipped. Root cause: `Testing/setupTests.ts` imported
   `@/shared/i18n` BEFORE installing the `MemoryStorage` polyfill, so
   `i18next-browser-languagedetector` probed jsdom's persistence-backed
   localStorage during init, hit the `--localstorage-file` path issue, and
   permanently cached `hasLocalStorageSupport = false`. All subsequent
   `cacheUserLanguage` calls no-oped. Fix: extract the storage swap into
   `Testing/install-storage.ts` and import it first in `setupTests.ts` so the
   polyfill is in place when i18n's detector runs its one-shot probe.

2. `src/features/library/hooks/useTreeState.ts` carried two
   `eslint-disable-next-line react-hooks/set-state-in-effect` directives that
   became unused once the rule stopped firing. Removed the directives (plus
   the stale "Wave 30 Task 3" rationale comment); the substantive effects are
   unchanged.

After: `npm test` 791/791 passed; `npm run lint` 0 errors / 6 warnings
(was 8); `npm run build` clean.
